tasks are now correctly executed sequentially again
